BFF-645 Split summary page into different widgets
